components/icm20948/requirements.txt
numpy-quaternion~=2023.0.2
open3d~=0.18.0
pyserial~=3.5